% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_wrapper.R
\name{categorical_gwas}
\alias{categorical_gwas}
\title{Run categorical GWAS}
\usage{
categorical_gwas(
  y,
  zz,
  kk = NULL,
  x0 = NULL,
  trait_type = c("ordinal", "nominal"),
  method = NULL,
  null_method = NULL,
  null_fit = NULL,
  vc = NULL,
  ids = NULL,
  outdir = NULL,
  maxiter = 100,
  minerr = 1e-08
)
}
\arguments{
\item{y}{Phenotype object.}

\item{zz}{Marker genotype matrix, markers in rows and individuals in columns.}

\item{kk}{Kinship matrix.}

\item{trait_type}{Either "nominal" or "ordinal".}

\item{method}{Character vector of methods.}
}
\value{
A list of GWAS results.
}
\description{
Run categorical GWAS
}
